mysql-connector-driver/requirements.txt
autopep8==1.4.4
mysql-connector==2.2.9
pycodestyle==2.5.0
